reading floats from MS5 and LSM

+++ User/Inc/L2/Barometer.h
@@ -5,8 +5,8 @@
 
 typedef struct
 {
-    int32_t pressure;    // Compensated pressure in mbar
-    int32_t temperature; // Compensated temperature in degrees Celsius
+    float pressure;    // Compensated pressure in mbar
+    float temperature; // Compensated temperature in degrees Celsius
 } MS5_compensated_values_t;
 
 void MS5_initialize(void);
